psifold/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py37_0
appnope=0.1.0=py37_0
astor=0.8.0=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
c-ares=1.15.0=h1de35cc_1001
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
cycler=0.10.0=py37_0
decorator=4.4.1=py_0
freetype=2.9.1=hb4e5f40_0
gast=0.2.2=py37_0
google-pasta=0.1.8=py_0
grpcio=1.16.1=py37h044775b_1
h5py=2.10.0=py37h3134771_0
hdf5=1.10.4=hfa1e0ec_0
importlib_metadata=1.5.0=py37_0
intel-openmp=2019.4=233
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.12.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.16.0=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.1=py37_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37h0a44026_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.11.4=hd9629dc_0
libsodium=1.0.16=h3efe00b_0
markdown=3.1.1=py37_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h9aa3819_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.0.15=py37h5e564d8_0
mkl_random=1.1.0=py37ha771720_0
more-itertools=8.2.0=py_0
ncurses=6.2=h0a44026_0
ninja=1.9.0=py37h04f5b5a_0
numpy=1.18.1=py37h7241aed_0
numpy-base=1.18.1=py37h6575580_1
openssl=1.1.1d=h1de35cc_4
opt_einsum=3.1.0=py_0
packaging=20.1=py_0
parso=0.6.1=py_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pip=20.0.2=py37_1
pluggy=0.13.1=py37_0
prompt_toolkit=3.0.3=py_0
protobuf=3.11.4=py37h0a44026_0
ptyprocess=0.6.0=py37_0
py=1.8.1=py_0
pygments=2.5.2=py_0
pyparsing=2.4.6=py_0
pytest=5.3.5=py37_0
python=3.7.6=h359304d_2
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.7_0
pyzmq=18.1.1=py37h0a44026_0
readline=7.0=h1de35cc_5
scipy=1.4.1=py37h9fa6033_0
setuptools=45.2.0=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=ha441bb4_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=mkl_py37hda344b4_0
tensorflow-base=2.0.0=mkl_py37h66b1bf0_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py37_1
tk=8.6.8=ha441bb4_0
tornado=6.0.3=py37h1de35cc_3
traitlets=4.3.3=py37_0
wcwidth=0.1.8=py_0
werkzeug=1.0.0=py_0
wheel=0.34.2=py37_0
wrapt=1.11.2=py37h1de35cc_0
xz=5.2.4=h1de35cc_4
zeromq=4.3.1=h0a44026_3
zipp=2.2.0=py_0
zlib=1.2.11=h1de35cc_3
